// src/services/proxy/generated/models/OrderItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * The Golden Book API
 * API for a book-selling business
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thegoldenbook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface OrderItem
 */
export interface OrderItem {
    /**
     * 
     * @type {number}
     * @memberof OrderItem
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof OrderItem
     */
    price?: number;
    /**
     * 
     * @type {number}
     * @memberof OrderItem
     */
    quantity?: number;
    /**
     * 
     * @type {number}
     * @memberof OrderItem
     */
    orderId?: number;
    /**
     * 
     * @type {number}
     * @memberof OrderItem
     */
    bookId?: number;
    /**
     * 
     * @type {string}
     * @memberof OrderItem
     */
    bookTitle?: string;
}

/**
 * Check if a given object implements the OrderItem interface.
 */
export function instanceOfOrderItem(value: object): value is OrderItem {
    return true;
}

export function OrderItemFromJSON(json: any): OrderItem {
    return OrderItemFromJSONTyped(json, false);
}

export function OrderItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrderItem {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'price': json['price'] == null ? undefined : json['price'],
        'quantity': json['quantity'] == null ? undefined : json['quantity'],
        'orderId': json['orderId'] == null ? undefined : json['orderId'],
        'bookId': json['bookId'] == null ? undefined : json['bookId'],
        'bookTitle': json['bookTitle'] == null ? undefined : json['bookTitle'],
    };
}

export function OrderItemToJSON(json: any): OrderItem {
    return OrderItemToJSONTyped(json, false);
}

export function OrderItemToJSONTyped(value?: OrderItem | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'price': value['price'],
        'quantity': value['quantity'],
        'orderId': value['orderId'],
        'bookId': value['bookId'],
        'bookTitle': value['bookTitle'],
    };
}